+++ internal/callbacker/grpc_callbacker_test.go
@@ -11,14 +11,14 @@ import (
 	"google.golang.org/grpc"
 	"google.golang.org/protobuf/types/known/emptypb"
 
+	"github.com/bitcoin-sv/arc/internal/callbacker"
 	"github.com/bitcoin-sv/arc/internal/callbacker/callbacker_api"
 	"github.com/bitcoin-sv/arc/internal/callbacker/mocks"
 	"github.com/bitcoin-sv/arc/internal/metamorph/metamorph_api"
 	"github.com/bitcoin-sv/arc/internal/metamorph/store"
-	"github.com/bitcoin-sv/arc/pkg/callbacker"
 )
 
-func TestSendCallback(t *testing.T) {
+func TestSendGRPCCallback(t *testing.T) {
 	tt := []struct {
 		name          string
 		expectedCalls int
@@ -65,13 +65,18 @@ func TestSendCallback(t *testing.T) {
 
 	for _, tc := range tt {
 		t.Run(tc.name, func(t *testing.T) {
+			// given
 			apiClient := &mocks.CallbackerAPIClientMock{
 				SendCallbackFunc: func(_ context.Context, _ *callbacker_api.SendCallbackRequest, _ ...grpc.CallOption) (*emptypb.Empty, error) {
 					return nil, nil
 				},
 			}
-			grpcCallbacker := callbacker.NewGrpcCallbacker(apiClient, slog.New(slog.NewJSONHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelInfo})))
-			grpcCallbacker.SendCallback(context.Background(), tc.data)
+			sut := callbacker.NewGrpcCallbacker(apiClient, slog.New(slog.NewJSONHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelInfo})))
+
+			// when
+			sut.SendCallback(context.Background(), tc.data)
+
+			// then
 			require.Equal(t, tc.expectedCalls, len(apiClient.SendCallbackCalls()))
 		})
 	}
